Revert "refactor: replace vite-string-plugin with vite's ?raw imports"
Keeps the diff focused on the test environment switch. The plugin change
was an unrelated cleanup.

The glob fixture still imports its data instead of reading it with
node:fs, since the browser has no fs, so ".txt" keeps a module
declaration.
