+++ package.json
@@ -108,6 +108,7 @@
     "svgo": "4.0.2",
     "typescript": "6.0.3",
     "typescript-eslint": "8.66.0",
+    "vite-string-plugin": "2.0.5",
     "vitest": "4.1.10",
     "vue-tsc": "3.3.9"
   }
