+++ vite.config.ts
@@ -1,4 +1,5 @@
 import {build, defineConfig} from 'vite';
+import {stringPlugin} from 'vite-string-plugin';
 import {sharedPlugins, vueDefines} from './tools/shared.ts';
 import {licensePlugin, wrap} from 'rolldown-license-plugin';
 import {readFileSync, writeFileSync, mkdirSync, unlinkSync, globSync} from 'node:fs';
@@ -86,6 +87,7 @@ function iifeBuildOpts({sourceFileName, write}: {sourceFileName: string, write?:
       rolldownOptions: {output: {entryFileNames: entryFileName}},
       ...(write === false && {write: false}),
     },
+    plugins: [stringPlugin()],
   });
 }
