+++ web_src/js/svg.ts
@@ -1,93 +1,93 @@
 import {parseDom, serializeXml} from './utils.ts';
 import {htmlRaw} from './utils/html.ts';
-import giteaDoubleChevronLeft from '../../public/assets/img/svg/gitea-double-chevron-left.svg?raw';
-import giteaDoubleChevronRight from '../../public/assets/img/svg/gitea-double-chevron-right.svg?raw';
-import giteaEmptyCheckbox from '../../public/assets/img/svg/gitea-empty-checkbox.svg?raw';
-import giteaExclamation from '../../public/assets/img/svg/gitea-exclamation.svg?raw';
-import giteaFavicon from '../../public/assets/img/favicon.svg?raw';
-import giteaRunning from '../../public/assets/img/svg/gitea-running.svg?raw';
-import octiconArchive from '../../public/assets/img/svg/octicon-archive.svg?raw';
-import octiconArrowLeft from '../../public/assets/img/svg/octicon-arrow-left.svg?raw';
-import octiconArrowSwitch from '../../public/assets/img/svg/octicon-arrow-switch.svg?raw';
-import octiconBlocked from '../../public/assets/img/svg/octicon-blocked.svg?raw';
-import octiconBold from '../../public/assets/img/svg/octicon-bold.svg?raw';
-import octiconCheck from '../../public/assets/img/svg/octicon-check.svg?raw';
-import octiconCheckbox from '../../public/assets/img/svg/octicon-checkbox.svg?raw';
-import octiconCheckCircleFill from '../../public/assets/img/svg/octicon-check-circle-fill.svg?raw';
-import octiconChevronDown from '../../public/assets/img/svg/octicon-chevron-down.svg?raw';
-import octiconChevronLeft from '../../public/assets/img/svg/octicon-chevron-left.svg?raw';
-import octiconChevronRight from '../../public/assets/img/svg/octicon-chevron-right.svg?raw';
-import octiconCircle from '../../public/assets/img/svg/octicon-circle.svg?raw';
-import octiconClock from '../../public/assets/img/svg/octicon-clock.svg?raw';
-import octiconCode from '../../public/assets/img/svg/octicon-code.svg?raw';
-import octiconColumns from '../../public/assets/img/svg/octicon-columns.svg?raw';
-import octiconCopy from '../../public/assets/img/svg/octicon-copy.svg?raw';
-import octiconDiffAdded from '../../public/assets/img/svg/octicon-diff-added.svg?raw';
-import octiconDiffModified from '../../public/assets/img/svg/octicon-diff-modified.svg?raw';
-import octiconDiffRemoved from '../../public/assets/img/svg/octicon-diff-removed.svg?raw';
-import octiconDiffRenamed from '../../public/assets/img/svg/octicon-diff-renamed.svg?raw';
-import octiconDotFill from '../../public/assets/img/svg/octicon-dot-fill.svg?raw';
-import octiconDownload from '../../public/assets/img/svg/octicon-download.svg?raw';
-import octiconEye from '../../public/assets/img/svg/octicon-eye.svg?raw';
-import octiconFile from '../../public/assets/img/svg/octicon-file.svg?raw';
-import octiconFileCode from '../../public/assets/img/svg/octicon-file-code.svg?raw';
-import octiconFileDirectoryFill from '../../public/assets/img/svg/octicon-file-directory-fill.svg?raw';
-import octiconFileDirectoryOpenFill from '../../public/assets/img/svg/octicon-file-directory-open-fill.svg?raw';
-import octiconFileRemoved from '../../public/assets/img/svg/octicon-file-removed.svg?raw';
-import octiconFileSubmodule from '../../public/assets/img/svg/octicon-file-submodule.svg?raw';
-import octiconFileSymlinkFile from '../../public/assets/img/svg/octicon-file-symlink-file.svg?raw';
-import octiconFilter from '../../public/assets/img/svg/octicon-filter.svg?raw';
-import octiconGear from '../../public/assets/img/svg/octicon-gear.svg?raw';
-import octiconGitBranch from '../../public/assets/img/svg/octicon-git-branch.svg?raw';
-import octiconGitCommit from '../../public/assets/img/svg/octicon-git-commit.svg?raw';
-import octiconGitMerge from '../../public/assets/img/svg/octicon-git-merge.svg?raw';
-import octiconGitPullRequest from '../../public/assets/img/svg/octicon-git-pull-request.svg?raw';
-import octiconGitPullRequestClosed from '../../public/assets/img/svg/octicon-git-pull-request-closed.svg?raw';
-import octiconGitPullRequestDraft from '../../public/assets/img/svg/octicon-git-pull-request-draft.svg?raw';
-import octiconGrabber from '../../public/assets/img/svg/octicon-grabber.svg?raw';
-import octiconHeading from '../../public/assets/img/svg/octicon-heading.svg?raw';
-import octiconHistory from '../../public/assets/img/svg/octicon-history.svg?raw';
-import octiconHorizontalRule from '../../public/assets/img/svg/octicon-horizontal-rule.svg?raw';
-import octiconHome from '../../public/assets/img/svg/octicon-home.svg?raw';
-import octiconImage from '../../public/assets/img/svg/octicon-image.svg?raw';
-import octiconInfo from '../../public/assets/img/svg/octicon-info.svg?raw';
-import octiconIssueClosed from '../../public/assets/img/svg/octicon-issue-closed.svg?raw';
-import octiconIssueOpened from '../../public/assets/img/svg/octicon-issue-opened.svg?raw';
-import octiconItalic from '../../public/assets/img/svg/octicon-italic.svg?raw';
-import octiconKebabHorizontal from '../../public/assets/img/svg/octicon-kebab-horizontal.svg?raw';
-import octiconLink from '../../public/assets/img/svg/octicon-link.svg?raw';
-import octiconListOrdered from '../../public/assets/img/svg/octicon-list-ordered.svg?raw';
-import octiconListUnordered from '../../public/assets/img/svg/octicon-list-unordered.svg?raw';
-import octiconLock from '../../public/assets/img/svg/octicon-lock.svg?raw';
-import octiconMeter from '../../public/assets/img/svg/octicon-meter.svg?raw';
-import octiconMilestone from '../../public/assets/img/svg/octicon-milestone.svg?raw';
-import octiconMirror from '../../public/assets/img/svg/octicon-mirror.svg?raw';
-import octiconOrganization from '../../public/assets/img/svg/octicon-organization.svg?raw';
-import octiconPlay from '../../public/assets/img/svg/octicon-play.svg?raw';
-import octiconPlus from '../../public/assets/img/svg/octicon-plus.svg?raw';
-import octiconProject from '../../public/assets/img/svg/octicon-project.svg?raw';
-import octiconQuote from '../../public/assets/img/svg/octicon-quote.svg?raw';
-import octiconRepo from '../../public/assets/img/svg/octicon-repo.svg?raw';
-import octiconRepoForked from '../../public/assets/img/svg/octicon-repo-forked.svg?raw';
-import octiconRepoTemplate from '../../public/assets/img/svg/octicon-repo-template.svg?raw';
-import octiconRss from '../../public/assets/img/svg/octicon-rss.svg?raw';
-import octiconScreenFull from '../../public/assets/img/svg/octicon-screen-full.svg?raw';
-import octiconSearch from '../../public/assets/img/svg/octicon-search.svg?raw';
-import octiconSidebarCollapse from '../../public/assets/img/svg/octicon-sidebar-collapse.svg?raw';
-import octiconSidebarExpand from '../../public/assets/img/svg/octicon-sidebar-expand.svg?raw';
-import octiconSkip from '../../public/assets/img/svg/octicon-skip.svg?raw';
-import octiconStar from '../../public/assets/img/svg/octicon-star.svg?raw';
-import octiconStop from '../../public/assets/img/svg/octicon-stop.svg?raw';
-import octiconStrikethrough from '../../public/assets/img/svg/octicon-strikethrough.svg?raw';
-import octiconSync from '../../public/assets/img/svg/octicon-sync.svg?raw';
-import octiconTable from '../../public/assets/img/svg/octicon-table.svg?raw';
-import octiconTag from '../../public/assets/img/svg/octicon-tag.svg?raw';
-import octiconTrash from '../../public/assets/img/svg/octicon-trash.svg?raw';
-import octiconTriangleDown from '../../public/assets/img/svg/octicon-triangle-down.svg?raw';
-import octiconX from '../../public/assets/img/svg/octicon-x.svg?raw';
-import octiconXCircleFill from '../../public/assets/img/svg/octicon-x-circle-fill.svg?raw';
-import octiconZoomIn from '../../public/assets/img/svg/octicon-zoom-in.svg?raw';
-import octiconZoomOut from '../../public/assets/img/svg/octicon-zoom-out.svg?raw';
+import giteaDoubleChevronLeft from '../../public/assets/img/svg/gitea-double-chevron-left.svg';
+import giteaDoubleChevronRight from '../../public/assets/img/svg/gitea-double-chevron-right.svg';
+import giteaEmptyCheckbox from '../../public/assets/img/svg/gitea-empty-checkbox.svg';
+import giteaExclamation from '../../public/assets/img/svg/gitea-exclamation.svg';
+import giteaFavicon from '../../public/assets/img/favicon.svg';
+import giteaRunning from '../../public/assets/img/svg/gitea-running.svg';
+import octiconArchive from '../../public/assets/img/svg/octicon-archive.svg';
+import octiconArrowLeft from '../../public/assets/img/svg/octicon-arrow-left.svg';
+import octiconArrowSwitch from '../../public/assets/img/svg/octicon-arrow-switch.svg';
+import octiconBlocked from '../../public/assets/img/svg/octicon-blocked.svg';
+import octiconBold from '../../public/assets/img/svg/octicon-bold.svg';
+import octiconCheck from '../../public/assets/img/svg/octicon-check.svg';
+import octiconCheckbox from '../../public/assets/img/svg/octicon-checkbox.svg';
+import octiconCheckCircleFill from '../../public/assets/img/svg/octicon-check-circle-fill.svg';
+import octiconChevronDown from '../../public/assets/img/svg/octicon-chevron-down.svg';
+import octiconChevronLeft from '../../public/assets/img/svg/octicon-chevron-left.svg';
+import octiconChevronRight from '../../public/assets/img/svg/octicon-chevron-right.svg';
+import octiconCircle from '../../public/assets/img/svg/octicon-circle.svg';
+import octiconClock from '../../public/assets/img/svg/octicon-clock.svg';
+import octiconCode from '../../public/assets/img/svg/octicon-code.svg';
+import octiconColumns from '../../public/assets/img/svg/octicon-columns.svg';
+import octiconCopy from '../../public/assets/img/svg/octicon-copy.svg';
+import octiconDiffAdded from '../../public/assets/img/svg/octicon-diff-added.svg';
+import octiconDiffModified from '../../public/assets/img/svg/octicon-diff-modified.svg';
+import octiconDiffRemoved from '../../public/assets/img/svg/octicon-diff-removed.svg';
+import octiconDiffRenamed from '../../public/assets/img/svg/octicon-diff-renamed.svg';
+import octiconDotFill from '../../public/assets/img/svg/octicon-dot-fill.svg';
+import octiconDownload from '../../public/assets/img/svg/octicon-download.svg';
+import octiconEye from '../../public/assets/img/svg/octicon-eye.svg';
+import octiconFile from '../../public/assets/img/svg/octicon-file.svg';
+import octiconFileCode from '../../public/assets/img/svg/octicon-file-code.svg';
+import octiconFileDirectoryFill from '../../public/assets/img/svg/octicon-file-directory-fill.svg';
+import octiconFileDirectoryOpenFill from '../../public/assets/img/svg/octicon-file-directory-open-fill.svg';
+import octiconFileRemoved from '../../public/assets/img/svg/octicon-file-removed.svg';
+import octiconFileSubmodule from '../../public/assets/img/svg/octicon-file-submodule.svg';
+import octiconFileSymlinkFile from '../../public/assets/img/svg/octicon-file-symlink-file.svg';
+import octiconFilter from '../../public/assets/img/svg/octicon-filter.svg';
+import octiconGear from '../../public/assets/img/svg/octicon-gear.svg';
+import octiconGitBranch from '../../public/assets/img/svg/octicon-git-branch.svg';
+import octiconGitCommit from '../../public/assets/img/svg/octicon-git-commit.svg';
+import octiconGitMerge from '../../public/assets/img/svg/octicon-git-merge.svg';
+import octiconGitPullRequest from '../../public/assets/img/svg/octicon-git-pull-request.svg';
+import octiconGitPullRequestClosed from '../../public/assets/img/svg/octicon-git-pull-request-closed.svg';
+import octiconGitPullRequestDraft from '../../public/assets/img/svg/octicon-git-pull-request-draft.svg';
+import octiconGrabber from '../../public/assets/img/svg/octicon-grabber.svg';
+import octiconHeading from '../../public/assets/img/svg/octicon-heading.svg';
+import octiconHistory from '../../public/assets/img/svg/octicon-history.svg';
+import octiconHorizontalRule from '../../public/assets/img/svg/octicon-horizontal-rule.svg';
+import octiconHome from '../../public/assets/img/svg/octicon-home.svg';
+import octiconImage from '../../public/assets/img/svg/octicon-image.svg';
+import octiconInfo from '../../public/assets/img/svg/octicon-info.svg';
+import octiconIssueClosed from '../../public/assets/img/svg/octicon-issue-closed.svg';
+import octiconIssueOpened from '../../public/assets/img/svg/octicon-issue-opened.svg';
+import octiconItalic from '../../public/assets/img/svg/octicon-italic.svg';
+import octiconKebabHorizontal from '../../public/assets/img/svg/octicon-kebab-horizontal.svg';
+import octiconLink from '../../public/assets/img/svg/octicon-link.svg';
+import octiconListOrdered from '../../public/assets/img/svg/octicon-list-ordered.svg';
+import octiconListUnordered from '../../public/assets/img/svg/octicon-list-unordered.svg';
+import octiconLock from '../../public/assets/img/svg/octicon-lock.svg';
+import octiconMeter from '../../public/assets/img/svg/octicon-meter.svg';
+import octiconMilestone from '../../public/assets/img/svg/octicon-milestone.svg';
+import octiconMirror from '../../public/assets/img/svg/octicon-mirror.svg';
+import octiconOrganization from '../../public/assets/img/svg/octicon-organization.svg';
+import octiconPlay from '../../public/assets/img/svg/octicon-play.svg';
+import octiconPlus from '../../public/assets/img/svg/octicon-plus.svg';
+import octiconProject from '../../public/assets/img/svg/octicon-project.svg';
+import octiconQuote from '../../public/assets/img/svg/octicon-quote.svg';
+import octiconRepo from '../../public/assets/img/svg/octicon-repo.svg';
+import octiconRepoForked from '../../public/assets/img/svg/octicon-repo-forked.svg';
+import octiconRepoTemplate from '../../public/assets/img/svg/octicon-repo-template.svg';
+import octiconRss from '../../public/assets/img/svg/octicon-rss.svg';
+import octiconScreenFull from '../../public/assets/img/svg/octicon-screen-full.svg';
+import octiconSearch from '../../public/assets/img/svg/octicon-search.svg';
+import octiconSidebarCollapse from '../../public/assets/img/svg/octicon-sidebar-collapse.svg';
+import octiconSidebarExpand from '../../public/assets/img/svg/octicon-sidebar-expand.svg';
+import octiconSkip from '../../public/assets/img/svg/octicon-skip.svg';
+import octiconStar from '../../public/assets/img/svg/octicon-star.svg';
+import octiconStop from '../../public/assets/img/svg/octicon-stop.svg';
+import octiconStrikethrough from '../../public/assets/img/svg/octicon-strikethrough.svg';
+import octiconSync from '../../public/assets/img/svg/octicon-sync.svg';
+import octiconTable from '../../public/assets/img/svg/octicon-table.svg';
+import octiconTag from '../../public/assets/img/svg/octicon-tag.svg';
+import octiconTrash from '../../public/assets/img/svg/octicon-trash.svg';
+import octiconTriangleDown from '../../public/assets/img/svg/octicon-triangle-down.svg';
+import octiconX from '../../public/assets/img/svg/octicon-x.svg';
+import octiconXCircleFill from '../../public/assets/img/svg/octicon-x-circle-fill.svg';
+import octiconZoomIn from '../../public/assets/img/svg/octicon-zoom-in.svg';
+import octiconZoomOut from '../../public/assets/img/svg/octicon-zoom-out.svg';
 
 const svgs = {
   'gitea-double-chevron-left': giteaDoubleChevronLeft,
